requirements.txt
tqdm
numpy
matplotlib
pandas
scikit_learn
pytest
torchvision
torchinfo
torchmetrics
einops
black>=22.6
coverage>=6
darglint>=1.8
isort>=5.10
pre-commit>=2.20
pytest-cov>=3
seaborn
SciencePlots
scikit-image
codecov
jupyter
loguru
opencv-python
py_irt
fastdup>=1.78 # 1.65
torch-lr-finder
faiss-gpu
cleanlab
lxml
imagehash
html5lib
openpyxl
pillow==9.0.0
umap-learn
statsmodels
krippendorff
wandb
transformers
pyod
